# Back-end/documention/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.3.8.5",
        "title": "HiddenWarrionKPI",
        "description": "Documention for HiddenWarrionKPI"
    },
    "paths": {
...
        "/api/group/create": {
            "post": {
                "operationId": "createGroup",
                "description": "create group",
                "tags": ["Group"],
                "parameters": [
                    { 
                        "type": "string",
                        "in": "header",
                        "description": "JWT token",
                        "name": "Authorization",
                        "required": true
                    },
                    { 
                        "type": "string",
                        "in": "formData",
                        "description": "Group name",
                        "name": "name",
                        "required": true
                    },
                    { 
                        "type": "string",
                        "in": "formData",
                        "description": "type (public/private)",
                        "name": "type",
                        "required": true
                    },
                    { 
                        "type": "string",
                        "in": "formData",
                        "description": "",
                        "name": "description",
                        "required": true
                    }
                    
                ],
                "responses": {
                    "200": {
                        "description": "success create group",
                        "schema": {
                            "type": "string",
                            "example": "Group created"
                        }
                    },
                    "401": {
                        "description": "Token is invalid",
                        "schema": {
                            "properties": {
                                "status": {
                                    "type": "number",
                                    "example": "401"
                                },
                                "errorStatus": {
                                    "type": "number",
                                    "example": "4012"
                                },
                                "message": {
                                    "type": "string",
                                    "example": "Token is invalid"
                                }
                            }
                        }
                    },
                    "401 - 2": {
                        "description": "No token provided",
                        "schema": {
                            "properties": {
                                "status": {
                                    "type": "number",
                                    "example": "401"
                                },
                                "errorStatus": {
                                    "type": "number",
                                    "example": "4014"
                                },
                                "message": {
                                    "type": "string",
                                    "example": "No token provided"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Group is created already",
                        "schema": {
                            "properties": {
                                "status": {
                                    "type": "number",
                                    "example": "409"
                                },
                                "errorStatus": {
                                    "type": "number",
                                    "example": "4091"
                                },
                                "message": {
                                    "type": "string",
                                    "example": "Group is created already"
                                }
                            }
                        }
                    }
                    
                }
            }
        },
...
        "/api/users/avatar": {
            "post": {
                "operationId": "getGroup",
                "description": "add avatar to user",
                "tags": ["Users"],
                "parameters": [
                    { 
                        "type": "string",
                        "in": "header",
                        "description": "ActionToken for user",
                        "name": "Authorization",
                        "required": true
                    },
                    { 
                        "type": "file",
                        "in": "formData",
                        "name": "avatar",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success added avatar to user",
                        "schema": {
                            "properties": {
                                "ETag": {
                                    "type": "string",
                                    "example": "c61fb4ac819deb0895d4dae551fa46cc"
                                },
                                "ServerSideEncryption": {
                                    "type": "string",
                                    "example": "AES256"
                                },
                                "Location": {
                                    "type": "string",
                                    "example": "https://api.example.s3.amazonaws.com/image/user/531ef8f7eb1ffea1828c9412/avatarhash.jpg"
                                },
                                "key": {
                                    "type": "string",
                                    "example": "/image/user/531ef8f7eb1ffea1828c9412/avatarhash.jpg"
                                },
                                "Bucket": {
                                    "type": "string",
                                    "example": "example-backet-name"
                                }
                            }
                        }
                    },
                    "413": {
                        "description": "File is to big size",
                        "schema": {
                            "properties": {
                                "status": {
                                    "type": "number",
                                    "example": "409"
                                },
                                "errorStatus": {
                                    "type": "number",
                                    "example": "4132"
                                },
                                "message": {
                                    "type": "string",
                                    "example": "File is to big"
                                }
                            }
                        }
                    },
                    "419": {
                        "description": "File is not available extension",
                        "schema": {
                            "properties": {
                                "status": {
                                    "type": "number",
                                    "example": "409"
                                },
                                "errorStatus": {
                                    "type": "number",
                                    "example": "4191"
                                },
                                "message": {
                                    "type": "string",
                                    "example": "File is not available extension"
                                }
                            }
                        }
                    }
                }
            }

        },
...
    }
}